+++ src/lib/components/ConfirmModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ConfirmModal';
