--- src/lib/components/modals/index.ts
@@ -1,3 +1,0 @@
-export * from './CommonModal';
-export * from './ConfirmModal';
-export * from './ModalButton';
